refactor: Create separate CheatsheetContent class
