+++ src/web/extension-web.ts
@@ -7,7 +7,7 @@
 
 import * as vscode from 'vscode';
 
-import { Cheatsheet } from '../cheatsheet';
+import { Cheatsheet } from '../cheatsheet-panel';
 import { DEBUG } from '../config';
 
 /**
